# documents/healthcare/api/openapi.yaml
openapi: 3.1.0
info:
  title: AllGuds Platform API
  version: "1.0.0"
  description: |
    Multi-tenant health, safety and compliance platform API.

    ## Tenancy
    Every request is scoped to exactly one tenant, taken from the `tenant_id` claim of the
    bearer token. `X-Tenant-ID` and the request subdomain are diagnostic hints only and are
    never used for authorization; a mismatch returns `403 TENANT_MISMATCH`.

    ## Records
    Business entities are stored generically and described by a per-tenant type registry.
    `/records` is the canonical route. Tenants also expose generated aliases per installed
    record type (`/patients`, `/appointments`); those resolve to the same handler with `type`
    fixed and are documented in the tenant-specific version of this spec.

    ## Conventions
    - Cursor pagination only. Offsets are not supported.
    - `If-Match` is required on record mutations; omitting it returns `428`.
    - `Idempotency-Key` is required on creates and sync pushes.
    - Clients MUST ignore unknown response fields and unknown enum values.
  contact:
    name: API Support
    url: https://docs.allguds.com
  license:
    name: Proprietary
servers:
  - url: https://api.allguds.com/v1
    description: Production (canonical)
  - url: https://{tenant}.allguds.com/api/v1
    description: Tenant alias. Routing only - never an authorization input.
    variables:
      tenant:
        default: demo
security:
  - bearerAuth: []
...
paths:
...
  /records:
    get:
      tags: [Records]
      summary: List records
      parameters:
        - $ref: '#/components/parameters/RecordType'
        - name: status
          in: query
          schema:
            type: string
        - name: updated_after
          in: query
          schema:
            type: string
            format: date-time
        - name: q
          in: query
          description: Full-text search over title and description
          schema:
            type: string
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - name: include_count
          in: query
          description: Expensive on large tenants. Off by default.
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: A page of records
          headers:
            RateLimit-Limit:
              $ref: '#/components/headers/RateLimitLimit'
            RateLimit-Remaining:
              $ref: '#/components/headers/RateLimitRemaining'
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Record'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
    post:
      tags: [Records]
      summary: Create a record
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecordInput'
      responses:
        '201':
          description: Created
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Location:
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Record'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationFailed'
  /records/{record_id}:
    parameters:
      - $ref: '#/components/parameters/RecordId'
    get:
      tags: [Records]
      summary: Get a record
      responses:
        '200':
          description: The record
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Record'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Records]
      summary: Partially update a record
      description: |
        JSON Merge Patch (RFC 7396). `null` clears a field. Requires `If-Match` carrying the
        record's current version.
      parameters:
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Updated
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Record'
        '404':
          $ref: '#/components/responses/NotFound'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '428':
          $ref: '#/components/responses/PreconditionRequired'
    delete:
      tags: [Records]
      summary: Soft-delete a record
      parameters:
        - $ref: '#/components/parameters/IfMatch'
      responses:
        '204':
          description: Deleted
        '404':
          $ref: '#/components/responses/NotFound'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
        '428':
          $ref: '#/components/responses/PreconditionRequired'
...
  /imports:
    post:
      tags: [Imports]
      summary: Start an import job
      description: |
        Requires `records:import`. Returns `202 Accepted` with a `Location` header naming the
        job; the work happens asynchronously (`database/07` Part B).

        **A dry run is the normal first step, and the default.** It stages, cleanses and
        validates every row and writes nothing to `records`, so a customer can iterate on the
        mapping against real validation output. Committing means submitting the same mapping
        again with `is_dry_run: false`.

        Rows are supplied inline. `source_file_id` is accepted for when `/files` exists, but a
        stored file cannot yet be read, so a job naming one is rejected rather than silently
        importing nothing.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImportInput'
      responses:
        '202':
          description: Accepted; poll the job for progress
          headers:
            Location:
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ImportJob'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '429':
          $ref: '#/components/responses/RateLimited'
  /imports/{import_job_id}:
    get:
      tags: [Imports]
      summary: Read an import job
      description: |
        Requires `records:import`. Counts are authoritative here rather than in the queue:
        job state lives in PostgreSQL, and Redis carries only the wake-up signal.
      parameters:
        - $ref: '#/components/parameters/ImportJobId'
      responses:
        '200':
          description: The job
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ImportJob'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /imports/{import_job_id}/errors:
    get:
      tags: [Imports]
      summary: Rows that failed validation
      description: |
        Requires `records:import`. Each entry carries the raw source row, so **this endpoint
        returns customer data and is PHI-bearing whenever the target type is**. It is audited
        as an access for that reason.
      parameters:
        - $ref: '#/components/parameters/ImportJobId'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: A page of row errors
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/ImportRowError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /imports/{import_job_id}/reverse:
    post:
      tags: [Imports]
      summary: Reverse an imported batch
      description: |
        Requires `records:import`. Deletes every record carrying this `import_job_id`, and the
        links between them.

        **Refuses once the tenant has edited any imported record.** `database/07` is explicit
        that reversal is safe only before the tenant starts working on the data; past that
        point it destroys their own work rather than undoing ours. A job that has already been
        reversed, or that never loaded anything, is also refused.
      parameters:
        - $ref: '#/components/parameters/ImportJobId'
      responses:
        '200':
          description: Reversed
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ImportJob'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: The batch can no longer be reversed safely
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Access token from `/auth/login`. One hour lifetime. The `tenant_id` claim is the sole
        authoritative tenant context.
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        Server-to-server key, format `hsc_live_<prefix>_<secret>`. Scoped to a subset of the
        creating user's permissions. No user identity is associated with key-authenticated
        requests.
  parameters:
    RecordId:
      name: record_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    RecordType:
      name: type
      in: query
      description: Record type code from the tenant's registry.
      schema:
        type: string
      example: patient
    Cursor:
      name: cursor
      in: query
      description: Opaque pagination cursor from `meta.page.next_cursor`. Offsets are not supported.
      schema:
        type: string
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 50
        minimum: 1
        maximum: 200
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: The record's current version, as returned in `ETag`.
      schema:
        type: string
      example: '"18422"'
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        Client-generated UUID. Replays return the original response with
        `Idempotency-Replayed: true`. Reusing a key with a different body is rejected.
      schema:
        type: string
        format: uuid
  headers:
    ETag:
      description: The record's current version.
      schema:
        type: string
    RateLimitLimit:
      description: Requests permitted in the current window.
      schema:
        type: integer
    RateLimitRemaining:
      description: Requests remaining in the current window.
      schema:
        type: integer
  responses:
    Unauthorized:
      description: |
        Authentication failed. `TOKEN_EXPIRED` means refresh and retry; `TOKEN_REVOKED` and
        `TOKEN_REUSE_DETECTED` mean log in again.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Forbidden:
      description: Authenticated but not permitted, or over a plan quota.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    NotFound:
      description: |
        No such resource in this tenant. Also returned when a resource exists in another
        tenant, so that existence is not disclosed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    ValidationFailed:
      description: Request understood but semantically invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    PreconditionFailed:
      description: '`If-Match` did not match the current version. Re-fetch and retry.'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    PreconditionRequired:
      description: '`If-Match` is required on this mutation and was absent.'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    RateLimited:
      description: Rate limit exceeded.
      headers:
        Retry-After:
          description: Seconds to wait before retrying.
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  schemas:
...
    Envelope:
      type: object
      required: [success, meta]
      properties:
        success:
          type: boolean
          const: true
        meta:
          $ref: '#/components/schemas/Meta'
...
    Record:
      type: object
      required: [record_id, record_type, version, created_at, updated_at]
      properties:
        record_id:
          type: string
          format: uuid
        record_type:
          type: string
        title:
          type: string
        description:
          type: string
        data:
          type: object
          description: Validated against the form version referenced by `form_version_id`.
          additionalProperties: true
        status:
          type: string
        workflow_state:
          type: string
        version:
          type: integer
          description: Increments on every update. Used for `If-Match` and sync conflicts.
        form_version_id:
          type: string
          format: uuid
        created_by:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    RecordInput:
      type: object
      required: [record_type]
      properties:
        record_type:
          type: string
        title:
          type: string
        description:
          type: string
        data:
          type: object
          additionalProperties: true
        status:
          type: string
        external_id:
          type: string
          description: Source-system key. Makes repeated imports idempotent.
...
    ImportInput:
      type: object
      required: [target_record_type, mappings, rows]
      properties:
        source_type:
          type: string
          enum: [csv, api, manual]
          default: api
        source_file_id:
          type: string
          format: uuid
          description: Reserved for when `/files` exists. Rejected until then.
        target_record_type:
          type: string
          maxLength: 100
          description: Must be a declared record type in this tenant's registry.
        is_dry_run:
          type: boolean
          default: true
          description: |
            Defaults to true. A dry run validates everything and writes no records, which is
            the loop a customer iterates on before committing.
        mapping_template_name:
          type: string
          maxLength: 150
          description: Save this mapping under a name for reuse on later jobs.
        mappings:
          type: array
          minItems: 1
          maxItems: 200
          items:
            $ref: '#/components/schemas/ImportFieldMapping'
        rows:
          type: array
          minItems: 1
          maxItems: 5000
          description: |
            Source rows, as flat objects keyed by source column. Bounded per request — a larger
            load is several jobs until `/files` can carry the source document.
          items:
            type: object
            additionalProperties: true
...